requirements.txt
tensorflow==2.6.2
pandas==1.3.5
numpy==1.19.2
scikit-learn==1.1.1
hydra-core==1.2.0
wandb==0.13.3
gpflow==2.5.2
matplotlib==3.5.0
plotly==5.10.0
tensorflow-probability==0.12.0
tensorflow-estimator==2.6.0
jupyter-core==4.10.0
